Finished MatchDetails and MatchLineUp XML files implementation. Simplified using statements. Temporarily removed the used of MatchRole enum since the documentation was incomplete and there were missing roles. Updated DownloadTask and download process to take advantage of Avalonia framework. Added Database object Schema and re-organized domain entities and mappings to reflect it. Added return value to InsertAsync. Added new translations. Re-done Initial migration. Finished DownloadView

+++ src/Hyperar.HPA.Common/Enums/AggressivenessLevel.cs
@@ -1,6 +1,6 @@
 ﻿namespace Hyperar.HPA.Common.Enums
 {
-    public enum AggressivenessLevel : uint
+    public enum AggressivenessLevel : byte
     {
         Tranquil = 0,
